// src/scripts/types/business.d.ts
export type Business =
  | "林業"
  | "海面漁業"
  | "定置網漁業又は海面魚類養殖業"
  | "金属鉱業、非金属鉱業又は石炭鉱業"
  | "石灰石鉱業又はドロマイト鉱業"
  | "原油又は天然ガス鉱業"
  | "採石業"
  | "その他の鉱業"
  | "水力発電施設、隧道等新設事業"
  | "道路新設事業"
  | "舗装工事業"
  | "鉄道又は軌道新設事業"
  | "建築事業"
  | "既設建築物設備工事業"
  | "機械装置の組立て又は据付けの事業"
  | "その他の建設事業"
  | "食料品製造業"
  | "繊維工業又は繊維製品製造業"
  | "木材又は木製品製造業"
  | "パルプ又は紙製造業"
  | "印刷又は製本業"
  | "化学工業"
  | "ガラス又はセメント製造業"
  | "コンクリート製造業"
  | "陶磁器製品製造業"
  | "その他の窯業又は土石製品製造業"
  | "金属精錬業"
  | "非鉄金属精錬業"
  | "金属材料品製造業"
  | "鋳物業"
  | "金属製品製造業又は金属加工業"
  | "洋食器、刃物、手工具又は一般金物製造業"
  | "めつき業"
  | "機械器具製造業"
  | "電気機械器具製造業"
  | "輸送用機械器具製造業"
  | "船舶製造又は修理業"
  | "計量器、光学機械、時計等製造業"
  | "貴金属製品、装身具、皮革製品等製造業"
  | "その他の製造業"
  | "交通運輸事業"
  | "貨物取扱事業"
  | "港湾貨物取扱事業"
  | "港湾荷役業"
  | "電気、ガス、水道又は熱供給の事業"
  | "農業又は海面漁業以外の漁業"
  | "清掃、火葬又は屠畜の事業"
  | "ビルメンテナンス業"
  | "倉庫業、警備業、消毒又は害虫駆除の事業又はゴルフ場の事業"
  | "通信業、放送業、新聞業又は出版業"
  | "卸売業・小売業、飲食店又は宿泊業"
  | "金融業、保険業又は不動産業"
  | "その他の各種事業"
  | "船舶所有者の事業";